refactor(condo): DOMA-9723 some fixes after review
