+++ apps/condo/domains/common/components/containers/BaseLayout/Header.tsx
@@ -4,7 +4,7 @@ import { OrganizationTypeType } from '@app/condo/schema'
 import { Layout } from 'antd'
 import get from 'lodash/get'
 import { useRouter } from 'next/router'
-import React, { useCallback, useMemo } from 'react'
+import React, { useCallback } from 'react'
 
 import { Menu } from '@open-condo/icons'
 import { useMutation } from '@open-condo/next/apollo'
@@ -17,11 +17,11 @@ import { Logo } from '@condo/domains/common/components/Logo'
 import { ResidentActions } from '@condo/domains/common/components/ResidentActions/ResidentActions'
 import { UserMessagesList } from '@condo/domains/notification/components/UserMessagesList'
 import { UserMessagesListContextProvider } from '@condo/domains/notification/contexts/UserMessagesListContext'
-import { OrganizationEmployeeRequests } from '@condo/domains/organization/components/OrganizationEmployeeRequests'
 import { InlineOrganizationSelect } from '@condo/domains/organization/components/OrganizationSelect'
 import { SBBOLIndicator } from '@condo/domains/organization/components/SBBOLIndicator'
 import { MANAGING_COMPANY_TYPE, SERVICE_PROVIDER_TYPE } from '@condo/domains/organization/constants/common'
 import { ACCEPT_OR_REJECT_ORGANIZATION_INVITE_BY_ID_MUTATION } from '@condo/domains/organization/gql'
+import { useOrganizationEmployeeRequests } from '@condo/domains/organization/hooks/useOrganizationEmployeeRequests'
 import { useOrganizationInvites } from '@condo/domains/organization/hooks/useOrganizationInvites'
 import { UserMenu } from '@condo/domains/user/components/UserMenu'
 
@@ -61,8 +61,8 @@ export const Header: React.FC<IHeaderProps> = (props) => {
         },
     })
 
-    const isEmployeeExist = useMemo(() => !!organization, [organization])
-    useOrganizationInvites(ORGANIZATION_TYPES, acceptOrReject, !isEmployeeExist)
+    useOrganizationInvites(ORGANIZATION_TYPES, acceptOrReject)
+    const { ChooseEmployeeRoleModal } = useOrganizationEmployeeRequests()
 
     const handleLogoClick = useCallback(() => {
         if (isAuthenticated) {
@@ -74,7 +74,7 @@ export const Header: React.FC<IHeaderProps> = (props) => {
 
     return (
         <UserMessagesListContextProvider>
-            {isEmployeeExist && <OrganizationEmployeeRequests />}
+            {ChooseEmployeeRoleModal}
             {
                 !breakpoints.TABLET_LARGE
                     ? (
