+++ apps/condo/domains/organization/components/OrganizationSelect.tsx
@@ -44,11 +44,11 @@ export const InlineOrganizationSelect: React.FC = () => {
     const AddOrganizationTitle = intl.formatMessage({ id: 'pages.organizations.CreateOrganizationButtonLabel' })
     const OkMessage = intl.formatMessage({ id: 'OK' })
     const EmployeeRequestDescription = intl.formatMessage({ id: 'organization.createOrganizationForm.request.description' })
-    const ChatInTelegramMessage = intl.formatMessage({ id: 'organization.createOrganizationForm.supportChat' })
+    const SupportChatMessage = intl.formatMessage({ id: 'organization.createOrganizationForm.supportChat' })
     const EmployeeRequestAlertDescription = intl.formatMessage({ id: 'organization.createOrganizationForm.request.alert.description' }, {
         chatBotLink: (
             <SecondaryLink target='_blank' href={HelpRequisites?.support_bot ? `https://t.me/${HelpRequisites.support_bot}` : '#'}>
-                {ChatInTelegramMessage}
+                {SupportChatMessage}
             </SecondaryLink>
         ),
     })
