+++ apps/condo/domains/organization/schema/OrganizationEmployeeRequest.js
@@ -115,7 +115,6 @@ const OrganizationEmployeeRequest = new GQLListSchema('OrganizationEmployeeReque
         userName: {
             schemaDoc: 'The name of the user who sent the request (For the organization employee because he does not have access to the user fields)',
             type: 'Text',
-            isRequired: true,
             access: {
                 read: true,
                 create: false,
@@ -126,7 +125,6 @@ const OrganizationEmployeeRequest = new GQLListSchema('OrganizationEmployeeReque
         userPhone: {
             schemaDoc: 'The phone of the user who sent the request (For the organization employee because he does not have access to the user fields)',
             type: 'Text',
-            isRequired: true,
             access: {
                 read: true,
                 create: false,
