+++ apps/condo/gql/operation.types.ts
@@ -381,7 +381,7 @@ export type GetActiveOrganizationEmployeeQueryVariables = Types.Exact<{
 }>;
 
 
-export type GetActiveOrganizationEmployeeQuery = { __typename?: 'Query', employees?: Array<{ __typename?: 'OrganizationEmployee', id: string, name?: string | null, email?: string | null, phone?: string | null, hasAllSpecializations?: boolean | null, isRejected?: boolean | null, isAccepted?: boolean | null, isBlocked?: boolean | null, position?: string | null, organization?: { __typename?: 'Organization', id: string, country?: Types.OrganizationCountryType | null, name?: string | null, type?: Types.OrganizationTypeType | null, tin?: string | null, features: Array<Types.OrganizationFeature>, statusTransitions?: any | null, importId?: string | null, importRemoteSystem?: string | null, meta?: any | null } | null, role?: { __typename?: 'OrganizationEmployeeRole', id: string, name?: string | null, nameNonLocalized?: string | null, isEditable?: boolean | null, isDefault?: boolean | null, statusTransitions?: any | null, ticketVisibilityType?: Types.OrganizationEmployeeRoleTicketVisibilityTypeType | null, canReadAnalytics?: boolean | null, canManageOrganization?: boolean | null, canManageCallRecords?: boolean | null, canDownloadCallRecords?: boolean | null, canReadEmployees?: boolean | null, canManageEmployees?: boolean | null, canInviteNewOrganizationEmployees?: boolean | null, canManageRoles?: boolean | null, canManageTicketPropertyHints?: boolean | null, canManageIntegrations?: boolean | null, canImportBillingReceipts?: boolean | null, canReadBillingReceipts?: boolean | null, canReadPayments?: boolean | null, canManageProperties?: boolean | null, canReadProperties?: boolean | null, canReadDocuments?: boolean | null, canManageDocuments?: boolean | null, canReadTickets?: boolean | null, canManageTickets?: boolean | null, canReadContacts?: boolean | null, canManageContacts?: boolean | null, canManageContactRoles?: boolean | null, canManageTicketComments?: boolean | null, canManagePropertyScopes?: boolean | null, canShareTickets?: boolean | null, canBeAssignedAsResponsible?: boolean | null, canBeAssignedAsExecutor?: boolean | null, canManageMeters?: boolean | null, canManageMeterReadings?: boolean | null, canManageBankAccounts?: boolean | null, canManageBankContractorAccounts?: boolean | null, canManageBankIntegrationAccountContexts?: boolean | null, canManageBankIntegrationOrganizationContexts?: boolean | null, canManageBankTransactions?: boolean | null, canManageBankAccountReports?: boolean | null, canManageBankAccountReportTasks?: boolean | null, canReadIncidents?: boolean | null, canManageIncidents?: boolean | null, canReadNewsItems?: boolean | null, canManageNewsItems?: boolean | null, canManageNewsItemTemplates?: boolean | null, canManageMobileFeatureConfigs?: boolean | null, canManageB2BApps?: boolean | null, canReadMeters?: boolean | null, canReadSettings?: boolean | null, canReadExternalReports?: boolean | null, canReadServices?: boolean | null, canReadCallRecords?: boolean | null, canReadInvoices?: boolean | null, canManageInvoices?: boolean | null, canReadMarketItems?: boolean | null, canManageMarketItems?: boolean | null, canManageMarketItemPrices?: boolean | null, canReadMarketItemPrices?: boolean | null, canReadMarketPriceScopes?: boolean | null, canManageMarketPriceScopes?: boolean | null, canReadMarketplace?: boolean | null, canManageMarketplace?: boolean | null, canReadPaymentsWithInvoices?: boolean | null, canReadTour?: boolean | null, canManageTour?: boolean | null, canReadMarketSetting?: boolean | null, canManageMarketSetting?: boolean | null, canManageTicketAutoAssignments?: boolean | null } | null } | null> | null };
+export type GetActiveOrganizationEmployeeQuery = { __typename?: 'Query', employees?: Array<{ __typename?: 'OrganizationEmployee', id: string, name?: string | null, email?: string | null, phone?: string | null, hasAllSpecializations?: boolean | null, isRejected?: boolean | null, isAccepted?: boolean | null, isBlocked?: boolean | null, position?: string | null, organization?: { __typename?: 'Organization', id: string, country?: Types.OrganizationCountryType | null, name?: string | null, type?: Types.OrganizationTypeType | null, tin?: string | null, features: Array<Types.OrganizationFeature>, statusTransitions?: any | null, importId?: string | null, importRemoteSystem?: string | null, meta?: any | null } | null, role?: { __typename?: 'OrganizationEmployeeRole', id: string, name?: string | null, nameNonLocalized?: string | null, isEditable?: boolean | null, isDefault?: boolean | null, statusTransitions?: any | null, ticketVisibilityType?: Types.OrganizationEmployeeRoleTicketVisibilityTypeType | null, canReadAnalytics?: boolean | null, canManageOrganization?: boolean | null, canManageOrganizationEmployeeRequests?: boolean | null, canManageCallRecords?: boolean | null, canDownloadCallRecords?: boolean | null, canReadEmployees?: boolean | null, canManageEmployees?: boolean | null, canInviteNewOrganizationEmployees?: boolean | null, canManageRoles?: boolean | null, canManageTicketPropertyHints?: boolean | null, canManageIntegrations?: boolean | null, canImportBillingReceipts?: boolean | null, canReadBillingReceipts?: boolean | null, canReadPayments?: boolean | null, canManageProperties?: boolean | null, canReadProperties?: boolean | null, canReadDocuments?: boolean | null, canManageDocuments?: boolean | null, canReadTickets?: boolean | null, canManageTickets?: boolean | null, canReadContacts?: boolean | null, canManageContacts?: boolean | null, canManageContactRoles?: boolean | null, canManageTicketComments?: boolean | null, canManagePropertyScopes?: boolean | null, canShareTickets?: boolean | null, canBeAssignedAsResponsible?: boolean | null, canBeAssignedAsExecutor?: boolean | null, canManageMeters?: boolean | null, canManageMeterReadings?: boolean | null, canManageBankAccounts?: boolean | null, canManageBankContractorAccounts?: boolean | null, canManageBankIntegrationAccountContexts?: boolean | null, canManageBankIntegrationOrganizationContexts?: boolean | null, canManageBankTransactions?: boolean | null, canManageBankAccountReports?: boolean | null, canManageBankAccountReportTasks?: boolean | null, canReadIncidents?: boolean | null, canManageIncidents?: boolean | null, canReadNewsItems?: boolean | null, canManageNewsItems?: boolean | null, canManageNewsItemTemplates?: boolean | null, canManageMobileFeatureConfigs?: boolean | null, canManageB2BApps?: boolean | null, canReadMeters?: boolean | null, canReadSettings?: boolean | null, canReadExternalReports?: boolean | null, canReadServices?: boolean | null, canReadCallRecords?: boolean | null, canReadInvoices?: boolean | null, canManageInvoices?: boolean | null, canReadMarketItems?: boolean | null, canManageMarketItems?: boolean | null, canManageMarketItemPrices?: boolean | null, canReadMarketItemPrices?: boolean | null, canReadMarketPriceScopes?: boolean | null, canManageMarketPriceScopes?: boolean | null, canReadMarketplace?: boolean | null, canManageMarketplace?: boolean | null, canReadPaymentsWithInvoices?: boolean | null, canReadTour?: boolean | null, canManageTour?: boolean | null, canReadMarketSetting?: boolean | null, canManageMarketSetting?: boolean | null, canManageTicketAutoAssignments?: boolean | null } | null } | null> | null };
 
 export type GetActualOrganizationEmployeesQueryVariables = Types.Exact<{
   userId: Types.Scalars['ID']['input'];
