+++ apps/condo/lang/en/en.json
@@ -454,10 +454,10 @@
   "api.organization.replaceOrganizationEmployeeRole.ORGANIZATION_NOT_FOUND": "Organization not found",
   "api.organization.replaceOrganizationEmployeeRole.ROLES_ARE_BEING_PROCESSED": "These roles are already being processed. Please try again a little later",
   "api.organization.replaceOrganizationEmployeeRole.ROLES_MUST_BE_DIFFERENT": "The new and old roles cannot be the same",
-  "api.organization.sendOrganizationEmployeeRequest.EMPLOYEE_ALREADY_ACCEPTED": "You have already been invited to this organization",
+  "api.organization.sendOrganizationEmployeeRequest.EMPLOYEE_ALREADY_ACCEPTED": "You are already a member of this organization",
   "api.organization.sendOrganizationEmployeeRequest.EMPLOYEE_INVITATION_ALREADY_SENT": "You are already invited to this organization",
-  "api.organization.sendOrganizationEmployeeRequest.REQUEST_NOT_PROCESSED": "A request has not been processed yet. Please wait for a decision on the request from the organization",
-  "api.organization.sendOrganizationEmployeeRequest.REQUEST_TO_ORGANIZATION_LIMIT_REACHED": "The number of requests to the organization has been exceeded",
+  "api.organization.sendOrganizationEmployeeRequest.REQUEST_NOT_PROCESSED": "The request has not been processed yet. Please wait for the organization's decision",
+  "api.organization.sendOrganizationEmployeeRequest.REQUEST_TO_ORGANIZATION_LIMIT_REACHED": "The limit for organization requests has been exceeded",
   "api.property.exportPropertiesToExcel.NOTHING_TO_EXPORT": "No properties are found to export for current organization",
   "api.property.property.SAME_ADDRESS": "Property with the same address already exists in selected organization",
   "api.propertyScope.exportPropertyScopesToExcel.NOTHING_TO_EXPORT": "No property scopes found for export",
@@ -1496,8 +1496,8 @@
   "notification.UserMessagesSettingModal.messageType.TICKET_COMMENT_CREATED.label": "About new ticket comments",
   "notification.UserMessagesSettingModal.messageType.TICKET_CREATED.label": "About new tickets",
   "notification.UserMessagesSettingModal.sound": "Sound",
-  "notification.UserMessagesSettingModal.sound.tooltip.disabled": "Notifications currently arrive without sound",
-  "notification.UserMessagesSettingModal.sound.tooltip.enabled": "Notifications currently arrive with sound",
+  "notification.UserMessagesSettingModal.sound.tooltip.disabled": "Sound is disabled for notifications",
+  "notification.UserMessagesSettingModal.sound.tooltip.enabled": "Sound is enabled for notifications",
   "notification.UserMessagesSettingModal.title": "What kind of notifications do you want to see?",
   "notification.messages.B2B_APP_MESSAGE_PUSH.push.title": "{title}",
   "notification.messages.B2C_APP_MESSAGE_PUSH.push.title": "{title}",
@@ -1611,8 +1611,8 @@
   "notification.messages.VOIP_INCOMING_CALL_MESSAGE.push.title": "{title}",
   "onboarding.tourStep.tooltip.ticket.message": "Make sure the information is correct and create an ticket",
   "onboarding.tourStep.tooltip.ticket.title": "Multiple fields filled automatically",
-  "organization.createOrganizationForm.cancelModal.message": "Access to the platform will only be available after creating an organization. If you leave this page, the organization data will not be saved.",
-  "organization.createOrganizationForm.cancelModal.title": "Changed your mind about creating an organization?",
+  "organization.createOrganizationForm.cancelModal.message": "You’ll need to create an organization to access the platform. Your progress will be lost if you leave this page.",
+  "organization.createOrganizationForm.cancelModal.title": "Cancel organization creation?",
   "organization.createOrganizationForm.field.name.description": "Enter the official name of the organization. It will be automatically copied into various documents.",
   "organization.createOrganizationForm.foundOrganizationsModal.sendRequest": "Yes, join",
   "organization.createOrganizationForm.foundOrganizationsModal.title": "Are you an employee and want to join the organization?",
