+++ apps/condo/pages/auth/organization.tsx
@@ -5,6 +5,7 @@ import React, { CSSProperties } from 'react'
 import { useIntl } from '@open-condo/next/intl'
 
 import { PageComponentType } from '@condo/domains/common/types'
+import { isSafeUrl } from '@condo/domains/common/utils/url.utils'
 import { CreateOrganizationPageContent } from '@condo/domains/organization/components/CreateOrganizationPageContent'
 import AuthLayout from '@condo/domains/user/components/containers/AuthLayout'
 import { WelcomeHeaderTitle } from '@condo/domains/user/components/UserWelcomeTitle'
@@ -12,7 +13,7 @@ import { WelcomeHeaderTitle } from '@condo/domains/user/components/UserWelcomeTi
 
 const ORGANIZATION_FORM_WRAPPER_STYLES: CSSProperties = { maxWidth: '350px' }
 
-const SignInPage: PageComponentType = () => {
+const CreateOrganizationPage: PageComponentType = () => {
     const intl = useIntl()
     const CreateOrganizationTitle = intl.formatMessage({ id: 'pages.organizations.CreateOrganizationModalTitle' })
 
@@ -27,7 +28,24 @@ const SignInPage: PageComponentType = () => {
         </>
     )
 }
-SignInPage.container = AuthLayout
-SignInPage.headerAction = <WelcomeHeaderTitle userType='staff'/>
+CreateOrganizationPage.container = AuthLayout
+CreateOrganizationPage.headerAction = <WelcomeHeaderTitle userType='staff'/>
+CreateOrganizationPage.getPrefetchedData = async ({ context, activeEmployee }) => {
+    const next = context?.query?.next
+    const redirectUrl = (next && !Array.isArray(next) && isSafeUrl(next)) ? next : '/'
 
-export default SignInPage
+    if (activeEmployee) {
+        return {
+            redirect: {
+                destination: redirectUrl,
+                permanent: false,
+            },
+        }
+    }
+
+    return {
+        props: {},
+    }
+}
+
+export default CreateOrganizationPage
